adding jest and babel-jest and first 4 tests
